Add database migration support using goose
Integrated goose for database migrations and updated the schema.sql file to a new migration format. This ensures automatic database migrations at startup, enhancing maintainability and consistency of the database state.

+++ pkg/services/auth_test.go
@@ -96,7 +96,7 @@ func TestAuthClient_DeletePasswordTokens(t *testing.T) {
 	count, err := c.Queries.Test_CountPasswordTokensByUser(context.Background(), usr.ID)
 
 	require.NoError(t, err)
-	assert.Equal(t, 0, count)
+	assert.Equal(t, int64(0), count)
 }
 
 func TestAuthClient_RandomToken(t *testing.T) {
